# trimbledrive/docs/apispec.yaml
...
openapi: 3.0.1
info:
  title: Trimble Drive API
  description: |-
    The service provides capabilities for managing files in the context of Connected Data Environment
    serving the Industry Clouds needs.
    A "file" is an unstructured blob of information with set of standard and custom metadata attributes.

    **PRIOR TO USING THIS API, YOU MUST AGREE TO OUR [TERMS OF USE](https://connect.trimble.com/terms_of_service.html)**

    # Main Capabilities

    * Managing hierarchical file structures (files organized in to hierarchies with folders)
        * The "file" is an unstructured BLOB content with the metadata end user usually expects from the familiar
    "file" concept
        * File hierarchies are organized as "file spaces" to provide logical isolation. The file spaces can be visible
    to end users or completely hidden and used by internal systems to manage unstructured data/files
    * Basic file and folder operations: create/upload, move/rename, delete, copy
        * hierarchical/recursive operations (e.g. delete folder) are implemented asynchronously
        * concurrency control on all operations including upsert, insert-only and update-only semantics
    * Strict versioning for files and folders. All changes are trackable and addressable. User is able to see the
    state of the full folder structure as it was at any moment of the history. User is able to refer to any specific
    version of the file.
...
    (even those not visible to the user).
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
security:
  - Bearer: [ ]
...
paths:
  /spaces:
    post:
      tags:
        - Filespace
      summary: Create Filespace
      description: "This API creates a new filespace. If identifier is provided as part of the request it will be assigned to new filespace on creation otherwise service will generate a new random identifier.\
                  \ By default only creator has access to all data in the filespace. <br>\
                  _Note that a special/root folder with the same name as filespace will be created internally during filespace creation, this folder acts as the parent folder for all the items in the filespace. Root folder details can be accessed just like any other folder by supplying the identifier, there will not be any parent for this folder._ "
      operationId: createFileSpace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFileSpaceRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: The filespace with the same identifier already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "201":
          description: Resource created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileSpaceDetails'
...
  /spaces/{spaceId}/folders:
    post:
      tags:
        - "Folders"
      summary: "Create Folder"
      description: "You can create a new folder using this action."
      operationId: "createFolder"
      parameters:
        - name: spaceId
          in: path
          description: filespace ID
          required: true
          schema:
            type: string
        - name: fields
          in: query
          description: "Option to specify fields to be returned in the response. The minimum set of required fields are  returned by default for optimum performance."
          required: false
          explode: false
          schema:
            type: array
            items:
              type: string
              enum:
                - name
                - size
                - parentFolderId
                - topActiveParentFolderId
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderRequest"
        required: true
      responses:
        "201":
          description: "Created"
          headers:
            ETag:
              description: "Weak ETag that can be used for the concurrency control."
              schema:
                type: "string"
            Location:
              description: "The location of a created resource."
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderDetailsResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /spaces/{spaceId}/folders/{folderId}:
    get:
      tags:
        - Folders
      summary: Get Folder Details
      description: "Get Folder details for a given folder identifier. By default, this operation returns `404 NotFound` in case of folder not present or if the resource is in deleted state, this default behavior can be modified with the `deleted` query parameter. <br/>\
                    To fetch a specific version information, optional query parameter `version` can be specified. Specific version information will be returned regardless of the deletion state."
      operationId: getFolderDetails
      parameters:
        - name: spaceId
          in: path
          description: filespace ID
          required: true
          schema:
            type: string
        - name: folderId
          in: path
          required: true
          schema:
            type: string
        - name: deleted
          in: query
          description: "If specified and true, resource is returned even if it is marked as deleted. Otherwise 404 error returned for deleted resource."
          required: false
          schema:
            type: boolean
            default: false
        - name: fields
          in: query
          description: "Option to specify fields to be returned in the response. The minimum set of required fields are  returned by default for optimum performance."
          required: false
          explode: false
          schema:
            type: array
            items:
              type: string
              enum:
                - name
                - size
                - parentFolderId
                - topActiveParentFolderId
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderDetailsResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - "Folders"
      summary: "Move/Rename Folder"
      description: "You can use this API to move/rename the Folder.\n\nTo move a folder,\
        \ update the `parentFolderId` attribute.\n\nThe API also supports **If-Match**\
        \ optional header to provide the version, so you can verify you are working\
        \ on the latest version.\n\nIf version provided in the **If-Match** header is not\
        \ latest, error Response with Response code **412** will be send back to\
        \ the caller.\n\nIf the **If-Match** header value\
        \ is *  or header not provided, actions will be performed on the latest version."
      operationId: "updateFolder"
      parameters:
        - name: spaceId
          in: path
          description: filespace ID
          required: true
          schema:
            type: string
        - name: "folderId"
          in: "path"
          description: "ID of the Folder."
          required: true
          schema:
            type: "string"
        - name: "If-Match"
          in: "header"
          schema:
            type: "string"
        - name: fields
          in: query
          description: "Option to specify fields to be returned in the response. The minimum set of required fields are  returned by default for optimum performance."
          required: false
          explode: false
          schema:
            type: array
            items:
              type: string
              enum:
                - name
                - size
                - parentFolderId
                - topActiveParentFolderId
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFolderRequest"
        required: true
      responses:
        "200":
          description: OK
          headers: { }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderDetailsResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - "Folders"
      summary: "Delete Folder"
      description: "You can use this API to mark the folder as deleted.\n\nThe API also supports **If-Match** optional header to provide\
        \ the version, so you can verify you are working on the latest version.\n\n\
        If version provided in the **If-Match** header is not latest, error Response\
        \ with Response code **412** will be send back to the caller.\
        \n\nIf the **If-Match** header value is *  or header not provided,\
        \ actions will be performed on the latest version."
      operationId: "deleteFolder"
      parameters:
        - name: spaceId
          in: path
          description: filespace ID
          required: true
          schema:
            type: string
        - name: "folderId"
          in: "path"
          description: "ID of the Folder."
          required: true
          schema:
            type: "string"
        - name: "If-Match"
          in: "header"
          schema:
            type: "string"
        - name: fields
          in: query
          description: "Option to specify fields to be returned in the response. The minimum set of required fields are  returned by default for optimum performance."
          required: false
          explode: false
          schema:
            type: array
            items:
              type: string
              enum:
                - name
                - size
                - parentFolderId
                - topActiveParentFolderId
      responses:
        "200":
          description: OK
          headers: { }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderDetailsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      title: "The error response."
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: "The error code. A non-empty language-independent string."
        message:
          type: string
          description: A human-readable string describing the error. The `Content-Language` header MUST contain the language code from [RFC5646](https://docs.oasis-open.org/odata/odata-json-format/v4.01/odata-json-format-v4.01.html#rfc5646) corresponding to the language in which the value for message is written.
    CreateFileSpaceRequest:
      type: object
      properties:
        id:
          type: string
          description: The identifier for a new file space can be given in the request.Only url-safe characters are allowed.The id has to be unique and cannot be updated after creation. If id is not specified service will generate unique id.
          maximum: 128
        description:
          type: string
          description: The description of the file space.
          maximum: 1024
        enforceSafeNames:
          type: boolean
          description: Option to manage file and folder naming restrictions of a file space. By default, naming restrictions will be enabled. Setting this as **`false`** during file space creation will turn OFF the naming restriction for the file space.
...
    FileSpaceDetails:
      type: object
      required:
        - id
        - rootId
        - clock
        - createdOn
        - createdBy
        - modifiedOn
        - modifiedBy
        - enforceSafeNames
      properties:
        id:
          type: string
          description: The globally unique identifier of the file space.
        rootId:
          type: string
          description: Root folder identifier.
        description:
          type: string
          description: The description of the file space.
        deleted:
          type: boolean
          description: Indicates whether the space is in the deleted state. If space is deleted, the content of the space is not accessible.
        createdOn:
          type: string
          format: date-time
          readOnly: true
        createdBy:
          type: string
          readOnly: true
        modifiedOn:
          type: string
          format: date-time
          readOnly: true
        modifiedBy:
          type: string
          readOnly: true
        changeToken:
          type: string
          description: The change token that represents a state of the space and all it's content (resources in the space) at specific moment of time when this resource has been returned.  Can be used for fetching the changes after this moment of time in incremental synchronisation scenarios by passing this token to `ListChanges` API. The change in the value of this token means something has been changed in the space.
        enforceSafeNames:
          type: boolean
          description: Option to manage file and folder naming restrictions of a file space. By default, naming restrictions will be enabled. Setting this as **`false`** during file space creation will turn OFF the naming restriction for the file space.
    CreateFolderRequest:
      title: "CreateFolderRequest"
      required:
        - "parentFolderId"
        - "name"
      type: "object"
      properties:
        id:
          type: string
          description: The identifier for a new folder can be given in the request. The id has to be unique in the file space and cannot be updated after creation. If id is not specified service will generate unique id.
          maxLength: 128
        name:
          type: "string"
          minLength: 1
          maxLength: 255
        parentFolderId:
          type: "string"
          minLength: 1
          maxLength: 128
    UpdateFolderRequest:
      title: "UpdateFolderRequest"
      type: "object"
      properties:
        name:
          type: "string"
          minLength: 1
          maxLength: 255
          description: "The new name."
        parentFolderId:
          type: "string"
          minLength: 1
          maxLength: 128
          description: "The new parent folder identifier"
    FolderDetailsResponse:
      type: object
      required:
        - id
        - version
        - type
        - spaceId
        - createdBy
        - modifiedBy
        - createdOn
        - modifiedOn
      properties:
        id:
          type: string
        version:
          type: string
          pattern: '^([1-9]\d*)\.0$'
        parentFolderId:
          type: string
          description: Folder identifier of the parent folder. Not present for root folder or for files not attached to any folder (attachments).
        name:
          type: string
        type:
          type: string
          enum:
            - FOLDER
        spaceId:
          type: string
        size:
          type: integer
          format: int64
        createdBy:
          type: string
        modifiedBy:
          type: string
        createdOn:
          type: string
          format: date-time
        modifiedOn:
          type: string
          format: date-time
        deleted:
          type: boolean
        topActiveParentFolderId:
          type : string
          description: Id of the topmost parent folder which is not in deleted state. Normally this will be equal to space `rootFolderId` for not deleted resources and equal to `null` (missing in the response) for deleted resources. If this property is not `null` and not equal to space `rootFolderId` then this resource belongs to orphan subtree with not deleted resources that is not connected to the space root folder.
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT